python/requirements.txt
autoflake
black
isort
pillow
reportlab
